Create FormTitle, TOSInput, and ConcernsSelector components
