+++ src/Components/Layout/Footer/Footer.css
@@ -1,5 +1,5 @@
 footer {
-  max-width: 600px;
+  width: 600px;
   margin-top: 3rem;
 }
