+++ src/Components/Pages/Process/Form/FormTitle.tsx
@@ -1,0 +1,8 @@
+export const FormTitle = () => {
+  return(
+    <h3 className="form-heading">
+      Paste, upload, or select your Terms of Service from a list of
+      popular services.
+    </h3>
+  )
+}
